📦 NEW: Refactor

+++ src/modules/categories/categories.service.ts
@@ -1,12 +1,18 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { PrismaService } from '../prisma/prisma.service';
+import { PostsService } from '../posts/posts.service';
 
 @Injectable()
 export class CategoriesService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    private postsService: PostsService,
+  ) {}
 
   async create(createCategoryDto: CreateCategoryDto) {
+    await this.postsService.findOne(createCategoryDto.post_id);
+
     const categoryExists = await this.prisma.category.findFirst({
       where: {
         name: createCategoryDto.name,
@@ -38,6 +44,12 @@ export class CategoriesService {
   }
 
   async remove(id: number) {
+    if ((await this.findOne(id)) === null) {
+      return new BadRequestException('Category not found', {
+        cause: new Error(),
+      });
+    }
+
     return await this.prisma.category.delete({
       where: {
         id,
